requirements.txt
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==21.2.0
